+++ examples/logical_enclosures.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -22,8 +22,8 @@
 ###
 
 from pprint import pprint
+
 from hpOneView.oneview_client import OneViewClient
-from hpOneView.exceptions import HPOneViewException
 from config_loader import try_load_from_file
 
 config = {
@@ -36,161 +36,137 @@ config = {
 
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
-
 oneview_client = OneViewClient(config)
-try:
-    # The valid enclosure URIs need to be inserted sorted by URI
-    # The number of enclosure URIs must be equal to the enclosure count in the enclosure group
-    options = dict(
-        enclosureUris=[],
-        enclosureGroupUri="",
-        forceInstallFirmware=False,
-        name="LogicalEnclosure2"
-    )
 
-    # Get enclosure group for creating logical enclosure
-    enclosure_groups = oneview_client.enclosure_groups.get_all()
-    first_enclosure_groups = enclosure_groups[0]
-    options["enclosureGroupUri"] = first_enclosure_groups["uri"]
-    enclosure_count = first_enclosure_groups["enclosureCount"]
+enclosure_groups = oneview_client.enclosure_groups
+enclosures = oneview_client.enclosures
+logical_enclosures = oneview_client.logical_enclosures
 
-    # Get enclosures
-    enclosures = oneview_client.enclosures.get_all()
-    enclosure_uris = []
-    for i in range(0, enclosure_count):
-        enclosure_uris.append(enclosures[i]["uri"])
-    options["enclosureUris"] = sorted(enclosure_uris)
+# The valid enclosure URIs need to be inserted sorted by URI
+# The number of enclosure URIs must be equal to the enclosure count in the enclosure group
+options = dict(
+    enclosureUris=[],
+    enclosureGroupUri="",
+    forceInstallFirmware=False,
+    name="LogicalEnclosure2"
+)
 
-    # Create a logical enclosure
-    # This method is only available on HPE Synergy.
-    logical_enclosure_created = oneview_client.logical_enclosures.create(options)
-    print("Created logical enclosure'%s' successfully.\n  uri = '%s'" % (
-        logical_enclosure_created['name'],
-        logical_enclosure_created['uri'])
-    )
-
-    # Delete the logical enclosure created
-    # This method is only available on HPE Synergy.
-    oneview_client.logical_enclosures.delete(logical_enclosure_created)
-    print("Delete logical enclosure")
-except HPOneViewException as e:
-    print(e.msg)
+# Get all logical enclosures
+print("Get all logical enclosures")
+logical_enclosures_all = logical_enclosures.get_all()
+for enc in logical_enclosures_all:
+    print('   %s' % enc['name'])
 
 # Get first logical enclosure
-logical_enclosure = oneview_client.logical_enclosures.get_all()[0]
-print("Found logical enclosure '{}' at\n   uri: '{}'".format(
-    logical_enclosure['name'], logical_enclosure['uri']))
+logical_enclosure_all = logical_enclosures.get_all()
+if logical_enclosure_all:
+    logical_enclosure = logical_enclosure_all[0]
+    print("Found logical enclosure '{}' at\n   uri: '{}'".format(
+        logical_enclosure['name'], logical_enclosure['uri']))
 
-# Update the logical enclosure name
-logical_enclosure_name = logical_enclosure['name']
-logical_enclosure['name'] = logical_enclosure_name + "-updated"
-print("Update the logical enclosure to have a name of '%s'" %
-      logical_enclosure['name'])
-logical_enclosure = oneview_client.logical_enclosures.update(logical_enclosure)
-print("   Done. uri: '%s', 'name': '%s'" %
-      (logical_enclosure['uri'], logical_enclosure['name']))
-print("Reset name")
-logical_enclosure['name'] = logical_enclosure_name
-logical_enclosure = oneview_client.logical_enclosures.update(logical_enclosure)
-print("   Done.")
+    # Get logical enclosure by uri
+    logical_enclosure = logical_enclosures.get_by_uri(logical_enclosure['uri'])
+    print("Got logical enclosure '{}' by\n   uri: '{}'".format(
+        logical_enclosure.data['name'], logical_enclosure.data['uri']))
 
-# Get logical enclosure by id
-try:
-    logical_enclosure_by_id = oneview_client.logical_enclosures.get(
-        "acb17b89-6724-4602-818a-1ee20ed4ec60")
-    print("Got logical enclosure '{}' by id: 'acb17b89-6724-4602-818a-1ee20ed4ec60'\n   uri: '{}'".format(
-        logical_enclosure_by_id['name'], logical_enclosure_by_id['uri']))
-except HPOneViewException as e:
-    print(e.msg)
-
-# Get logical enclosure by uri
-logical_enclosure = oneview_client.logical_enclosures.get(
-    logical_enclosure['uri'])
-print("Got logical enclosure '{}' by\n   uri: '{}'".format(
-    logical_enclosure['name'], logical_enclosure['uri']))
-pprint(logical_enclosure)
-
-# Get logical enclosure by name
-try:
-    logical_enclosure_by_name = oneview_client.logical_enclosures.get_by('name', logical_enclosure['name'])[0]
-    print("Got logical enclosure by name '{name}'\n   uri: '{uri}'".format(**logical_enclosure_by_name))
-except HPOneViewException as e:
-    print(e.msg)
 # Get Logical Enclosure by scope_uris
-if oneview_client.api_version == 600:
-    le_by_scope_uris = oneview_client.logical_enclosures.get_all(scope_uris="\"'/rest/scopes/cd237b60-09e2-45c4-829e-082e318a6d2a'\"")
+if oneview_client.api_version >= 600:
+    le_by_scope_uris = logical_enclosures.get_all(scope_uris="\"'/rest/scopes/cd237b60-09e2-45c4-829e-082e318a6d2a'\"")
     if len(le_by_scope_uris) > 0:
         print("Got Logical Enclosure by scope_uris: '%s'.\n  uri = '%s'" % (le_by_scope_uris[0]['name'], le_by_scope_uris[0]['uri']))
         pprint(le_by_scope_uris)
     else:
         print("No Logical Enclosure found by scope_uris")
 
+# Get Logical Enclosure by name
+logical_enclosure = logical_enclosures.get_by_name(options["name"])
+if not logical_enclosure:
+    # Get enclosure group for creating logical enclosure
+    enclosure_groups_all = enclosure_groups.get_all()
+    first_enclosure_groups = enclosure_groups_all[0]
+    options["enclosureGroupUri"] = first_enclosure_groups["uri"]
+    enclosure_count = first_enclosure_groups["enclosureCount"]
+
+    # Get enclosures
+    enclosures_all = enclosures.get_all()
+    enclosure_uris = []
+    for i in range(0, enclosure_count):
+        enclosure_uris.append(enclosures_all[i]["uri"])
+    options["enclosureUris"] = sorted(enclosure_uris)
+
+    # Create a logical enclosure
+    # This method is only available on HPE Synergy.
+    logical_enclosure = logical_enclosures.create(options)
+    print("Created logical enclosure'%s' successfully.\n  uri = '%s'" % (
+        logical_enclosure['name'],
+        logical_enclosure['uri'])
+    )
+
+# Update the logical enclosure name
+print("Update the logical enclosure to have a name of '%s'" %
+      options["name"])
+resource = logical_enclosure.data.copy()
+previous_name = resource["name"]
+resource["name"] = resource["name"] + "-Renamed"
+logical_enclosure.update(resource)
+print("   Done. uri: '%s', 'name': '%s'" %
+      (logical_enclosure.data['uri'], logical_enclosure.data['name']))
+
+print("Reset name")
+resource["name"] = previous_name
+logical_enclosure.update(resource)
+print("   Done. uri: '%s', 'name': '%s'" %
+      (logical_enclosure.data['uri'],
+       logical_enclosure.data['name']))
+
 # Update configuration
 print("Reapply the appliance's configuration to the logical enclosure")
-try:
-    oneview_client.logical_enclosures.update_configuration(
-        logical_enclosure['uri'])
-    print("   Done.")
-except HPOneViewException as e:
-    print(e.msg)
+logical_enclosure.update_configuration()
+print("   Done.")
 
-# update and get script
+# Update and get script
 print("Update script")
-try:
-    script = "# TEST COMMAND"
-    logical_enclosure = oneview_client.logical_enclosures.update_script(
-        logical_enclosure['uri'], script)
-    print("   updated script: '{}'".format(
-        oneview_client.logical_enclosures.get_script(logical_enclosure['uri'])))
-except HPOneViewException as e:
-    print("  %s" % e.msg)
+script = "# TEST COMMAND"
+logical_enclosure_updated = logical_enclosure.update_script(
+    logical_enclosure.data['uri'], script)
+print("   updated script: '{}'".format(
+    logical_enclosure.get_script()))
 
-# create support dumps
+# Create support dumps
 print("Generate support dump")
 info = {
     "errorCode": "MyDump16",
     "encrypt": True,
     "excludeApplianceDump": False
 }
-support_dump = oneview_client.logical_enclosures.generate_support_dump(
-    info, logical_enclosure['uri'])
+support_dump = logical_enclosure.generate_support_dump(info)
 print("   Done")
 
 # update from group
 print("Update from group")
-try:
-    logical_enclosure = oneview_client.logical_enclosures.update_from_group(
-        logical_enclosure['uri'])
-    print("   Done")
-except HPOneViewException as e:
-    print("  %s" % e.msg)
-
-# Get all logical enclosures
-print("Get all logical enclosures")
-logical_enclosures = oneview_client.logical_enclosures.get_all()
-for enc in logical_enclosures:
-    print('   %s' % enc['name'])
+logical_enclosure_updated = logical_enclosure.update_from_group()
+print("   Done")
 
 if oneview_client.api_version >= 300:
-
-    if logical_enclosures:
-
+    if logical_enclosure:
         print("Update firmware for a logical enclosure with the logical-interconnect validation set as true.")
 
-        logical_enclosure = logical_enclosures[0]
-
-        logical_enclosure_updated = oneview_client.logical_enclosures.patch(
-            id_or_uri=logical_enclosure["uri"],
+        logical_enclosure_updated = logical_enclosure.patch(
             operation="replace",
             path="/firmware",
             value={
-                "firmwareBaselineUri": "/rest/firmware-drivers/SPPgen9snap6_2016_0405_87",
+                "firmwareBaselineUri": "/rest/firmware-drivers/SPP_2018_06_20180709_for_HPE_Synergy_Z7550-96524",
                 "firmwareUpdateOn": "EnclosureOnly",
                 "forceInstallFirmware": "true",
                 "validateIfLIFirmwareUpdateIsNonDisruptive": "true",
                 "logicalInterconnectUpdateMode": "Orchestrated",
                 "updateFirmwareOnUnmanagedInterconnect": "true"
-            }
+            },
+            custom_headers={"if-Match": "*"}
         )
+        pprint(logical_enclosure_updated.data)
 
-        pprint(logical_enclosure_updated)
+# Delete the logical enclosure created
+# This method is only available on HPE Synergy.
+logical_enclosure.delete()
+print("Delete logical enclosure")