+++ hpOneView/resources/servers/server_hardware_types.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -27,13 +27,14 @@ from __future__ import division
 from __future__ import absolute_import
 from future import standard_library
 
+
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import Resource, ensure_resource_client
 
 
-class ServerHardwareTypes(object):
+class ServerHardwareTypes(Resource):
     """
     The server hardware types resource is a representation/abstraction of a physical server managed by the appliance.
     It defines capabilities and settings that can be used in a server profile.
@@ -41,93 +42,23 @@ class ServerHardwareTypes(object):
     """
     URI = '/rest/server-hardware-types'
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(ServerHardwareTypes, self).__init__(connection, data)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets the list of server hardware type resources defined on the appliance.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: A list of server hardware types.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get(self, id_or_uri):
-        """
-        Gets the server hardware type resource with the specified ID or URI.
-
-        Args:
-            id_or_uri: Can be either the server hardware type ID or URI.
-
-        Returns:
-            dict: The logical interconnect group.
-        """
-        return self._client.get(id_or_uri)
-
-    def update(self, resource, uri=None, timeout=-1):
+    @ensure_resource_client
+    def update(self, data, timeout=-1, force=False):
         """
         Updates one or more attributes for a server hardware type resource.
-
         Args:
-            resource (dict): Object to update.
+            data (dict): Object to update.
             timeout:
                 Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
                 in OneView; it just stops waiting for its completion.
-
+            force: Flag to force the operation.
         Returns:
             dict: Updated server hardware type.
-
         """
-        return self._client.update(resource, uri, timeout=timeout)
+        uri = self.data["uri"]
+        self.data = self._helper.update(data, uri=uri, timeout=timeout, force=force)
 
-    def delete(self, resource, force=False, timeout=-1):
-        """
-        Removes the server hardware type with the specified URI. A server hardware type cannot be deleted if it
-        is associated with a server hardware or server profile resource.
-
-        Args:
-            resource (dict): object to delete
-            force (bool):
-                 If set to true, the operation completes despite any problems with
-                 network connectivity or errors on the resource itself. The default is false.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView; it just stops waiting for its completion.
-
-        Returns:
-            bool: Indicates whether the resource was successfully deleted.
-
-        """
-        return self._client.delete(resource, force=force, timeout=timeout)
-
-    def get_by(self, field, value):
-        """
-        Gets all server hardware types that match the filter.
-
-        The search is case-insensitive.
-
-        Args:
-            field: Field name to filter.
-            value: Value to filter.
-
-        Returns:
-            list: A list of server hardware types.
-        """
-        return self._client.get_by(field, value)
+        return self