+++ tests/unit/resources/networking/test_sas_interconnects.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -26,18 +26,18 @@ import unittest
 import mock
 
 from hpOneView.resources.networking.sas_interconnects import SasInterconnects
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import ResourceHelper, ResourcePatchMixin
 
 
 class SasInterconnectsTest(unittest.TestCase):
 
-    SAS_INTERCONNECT_URI = '/rest/sas-interconnects/3518be0e-17c1-4189-8f81-83f3724f6155'
-
     def setUp(self):
         self.host = '127.0.0.1'
         self._sas_interconnects = SasInterconnects(None)
+        self.uri = '/rest/sas-interconnects/3518be0e-17c1-4189-8f81-83f3724f6155'
+        self._sas_interconnects.data = {"uri": self.uri}
 
-    @mock.patch.object(ResourceClient, 'get_all')
+    @mock.patch.object(ResourceHelper, 'get_all')
     def test_get_all_called_once(self, mock_get_all):
         filter = 'name=TestName'
         sort = 'name:ascending'
@@ -45,43 +45,26 @@ class SasInterconnectsTest(unittest.TestCase):
         self._sas_interconnects.get_all(2, 500, filter=filter, sort=sort)
         mock_get_all.assert_called_once_with(start=2, count=500, filter=filter, sort=sort, query='', view='', fields='')
 
-    @mock.patch.object(ResourceClient, 'get_by')
-    def test_get_by_called_once(self, mock_get_by):
-        sas_interconnect_name = "0000A66103, interconnect 4"
-        self._sas_interconnects.get_by('name', sas_interconnect_name)
-        mock_get_by.assert_called_once_with('name', sas_interconnect_name)
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_called_once(self, mock_get):
-        sas_interconnect_id = '3518be0e-17c1-4189-8f81-83f3724f6155'
-        self._sas_interconnects.get(id_or_uri=sas_interconnect_id)
-        mock_get.assert_called_once_with(sas_interconnect_id)
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_with_uri_called_once(self, mock_get):
-        self._sas_interconnects.get(id_or_uri=self.SAS_INTERCONNECT_URI)
-        mock_get.assert_called_once_with(self.SAS_INTERCONNECT_URI)
-
-    @mock.patch.object(ResourceClient, 'patch')
-    def test_patch_called_once(self, mock_patch):
+    @mock.patch.object(ResourcePatchMixin, 'patch_request')
+    def test_patch_called_once(self, mock_patch_request):
         args = dict(
-            id_or_uri=self.SAS_INTERCONNECT_URI,
             operation='replace',
             path='/deviceResetState',
             value='Reset',
         )
 
         self._sas_interconnects.patch(**args)
-        mock_patch.assert_called_once_with(timeout=-1, **args)
+        mock_patch_request.assert_called_once_with(self.uri,
+                                                   body=[{'path': '/deviceResetState',
+                                                          'op': 'replace',
+                                                          'value': 'Reset'}],
+                                                   custom_headers=None, timeout=-1)
 
-    @mock.patch.object(ResourceClient, 'build_uri')
-    @mock.patch.object(ResourceClient, 'update')
-    def test_refresh_state_called_once(self, mock_update, mock_build_uri):
+    @mock.patch.object(ResourceHelper, 'update')
+    def test_refresh_state_called_once(self, mock_update):
         configuration = dict(refreshState="RefreshPending")
-        expected_uri = self.SAS_INTERCONNECT_URI + "/refreshState"
+        expected_uri = "{}/refreshState".format(self.uri)
 
-        mock_build_uri.return_value = self.SAS_INTERCONNECT_URI
-        self._sas_interconnects.refresh_state(id_or_uri=self.SAS_INTERCONNECT_URI, configuration=configuration)
+        self._sas_interconnects.refresh_state(configuration=configuration)
 
-        mock_build_uri.assert_called_once_with(self.SAS_INTERCONNECT_URI)
-        mock_update.assert_called_once_with(uri=expected_uri, resource=configuration)
+        mock_update.assert_called_once_with(resource=configuration, uri=expected_uri)